Añadir funciones para realizar pedidos y reservas. Tambien restaurar contraseña de usuario

+++ client/src/components/user/Logout.jsx
@@ -1,17 +1,28 @@
 import { useNavigate } from "react-router-dom";
 import useAuth from "../../hooks/useAuth";
 import { useEffect } from "react";
+import Swal from "sweetalert2";
 
 const Logout = () => {
-  const { setAuth } = useAuth();
+  const { setAuth, authUser } = useAuth();
   const navigate = useNavigate();
+
   useEffect(() => {
-    localStorage.clear();
-    setAuth(null);
-    navigate("/");
-    window.location.reload();
-  });
-  return <h1>Cerrando sesión...</h1>;
+    Swal.fire({
+      title: "Cerrando sesión...",
+      timer: 2000,
+      timerProgressBar: true,
+      didOpen: () => {
+        Swal.showLoading();
+      },
+    }).then(() => {
+      localStorage.clear();
+      setAuth({});
+      authUser();
+      navigate("/");
+    });
+  }, [setAuth, authUser, navigate]);
+  return null;
 };
 
 export default Logout;
